Add tsconfig
