+++ tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "compilerOptions": {
+    "lib": ["deno.ns", "dom"]
+  }
+}
